+++ include/JSystem/JDrama/JDRLighting.hpp
@@ -32,7 +32,7 @@ public:
 	}
 
 	virtual void load(JSUMemoryInputStream&);
-	virtual void perform(unsigned long, TGraphics*);
+	virtual void perform(u32, TGraphics*);
 
 public:
 	/* 0x10 */ s32 mLightInfoCount;
